+++ CHANGELOG.md
@@ -15,9 +15,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Option to show event location in calendar
 - Finnish translation for "Feels" and "Weeks"
 - Russian translation for “Feels”
+- Calendar module: added `nextDaysRelative` config option
+- Add `broadcastPastEvents` config option for calendars to include events from the past `maximumNumberOfDays` in event broadcasts
 
 ### Updated
 - English translation for "Feels" to "Feels like"
+- Fixed the example calender url in `config.js.sample`
+- Update `ical.js` to solve various calendar issues.
 
 ### Fixed
 - Handle SIGTERM messages